bpf: srv6: move SRv6 map structs to srv6.h
A bit more untangling of the big common.h header.
