+++ bpf/lib/common.h
@@ -155,30 +155,6 @@ struct auth_info {
 	__u64       expiration;
 };
 
-struct srv6_vrf_key4 {
-	struct bpf_lpm_trie_key lpm;
-	__u32 src_ip;
-	__u32 dst_cidr;
-};
-
-struct srv6_vrf_key6 {
-	struct bpf_lpm_trie_key lpm;
-	union v6addr src_ip;
-	union v6addr dst_cidr;
-};
-
-struct srv6_policy_key4 {
-	struct bpf_lpm_trie_key lpm;
-	__u32 vrf_id;
-	__u32 dst_cidr;
-};
-
-struct srv6_policy_key6 {
-	struct bpf_lpm_trie_key lpm;
-	__u32 vrf_id;
-	union v6addr dst_cidr;
-};
-
 #ifndef BPF_F_PSEUDO_HDR
 # define BPF_F_PSEUDO_HDR                (1ULL << 4)
 #endif
