+++ bpf/lib/srv6.h
@@ -6,7 +6,32 @@
 #include "lib/common.h"
 #include "lib/drop.h"
 #include "lib/identity.h"
+#include "lib/ipv6_core.h"
 #include "lib/tailcall.h"
+
+struct srv6_vrf_key4 {
+	struct bpf_lpm_trie_key lpm;
+	__u32 src_ip;
+	__u32 dst_cidr;
+};
+
+struct srv6_vrf_key6 {
+	struct bpf_lpm_trie_key lpm;
+	union v6addr src_ip;
+	union v6addr dst_cidr;
+};
+
+struct srv6_policy_key4 {
+	struct bpf_lpm_trie_key lpm;
+	__u32 vrf_id;
+	__u32 dst_cidr;
+};
+
+struct srv6_policy_key6 {
+	struct bpf_lpm_trie_key lpm;
+	__u32 vrf_id;
+	union v6addr dst_cidr;
+};
 
 struct {
 	__uint(type, BPF_MAP_TYPE_LPM_TRIE);
